Moved removeBuildDirectory into the Builder.
This required to split Builder into two classes to keep the the number of methods low.
The BaseBuilder implements the core features while the Builder adds the delegating methods.
