+++ PHPCI/BaseBuilder.php
@@ -1,0 +1,340 @@
+<?php
+/**
+ * PHPCI - Continuous Integration for PHP
+ *
+ * @copyright    Copyright 2014, Block 8 Limited.
+ * @license      https://github.com/Block8/PHPCI/blob/master/LICENSE.md
+ * @link         https://www.phptesting.org/
+ */
+
+namespace PHPCI;
+
+use b8\Config;
+use b8\Store\Factory;
+use DateTime;
+use Exception;
+use PHPCI\Helper\BuildInterpolator;
+use PHPCI\Helper\CommandExecutor;
+use PHPCI\Helper\Lang;
+use PHPCI\Helper\MailerFactory;
+use PHPCI\Logging\BuildLogger;
+use PHPCI\Model\Build;
+use PHPCI\Plugin\Util\Executor;
+use PHPCI\Plugin\Util\Factory as PluginFactory;
+use PHPCI\Store\BuildStore;
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\LoggerInterface;
+
+/**
+ * PHPCI Build Runner
+ *
+ * Implements the base Builder features.
+ *
+ * @author   Adirelle <adirelle@gmail.com>
+ */
+class BaseBuilder implements LoggerAwareInterface
+{
+    /**
+     * @var string
+     */
+    public $buildPath;
+
+    /**
+     * @var string[]
+     */
+    public $ignore = array();
+
+    /**
+     * @var bool
+     */
+    public $quiet = false;
+
+    /**
+     * @var bool
+     */
+    protected $verbose = true;
+
+    /**
+     * @var Build
+     */
+    protected $build;
+
+    /**
+     * @var LoggerInterface
+     */
+    protected $logger;
+
+    /**
+     * @var array
+     */
+    protected $config;
+
+    /**
+     * @var BuildInterpolator
+     */
+    protected $interpolator;
+
+    /**
+     * @var BuildStore
+     */
+    protected $store;
+
+    /**
+     * @var Executor
+     *
+     */
+    protected $pluginExecutor;
+
+    /**
+     * @var CommandExecutor
+     */
+    protected $commandExecutor;
+
+    /**
+     * @var BuildLogger
+     */
+    protected $buildLogger;
+
+    /**
+     * Set up the builder.
+     *
+     * @param Build $build
+     * @param LoggerInterface $logger
+     */
+    public function __construct(Build $build, LoggerInterface $logger = null)
+    {
+        $this->build = $build;
+        $this->store = Factory::getStore('Build');
+
+        $this->buildLogger = new BuildLogger($logger, $build);
+
+        $pluginFactory = $this->buildPluginFactory($build);
+        $pluginFactory->addConfigFromFile(PHPCI_DIR . "/pluginconfig.php");
+        $this->pluginExecutor = new Executor($pluginFactory, $this->buildLogger);
+
+        $executorClass = 'PHPCI\Helper\UnixCommandExecutor';
+        if (IS_WIN) {
+            $executorClass = 'PHPCI\Helper\WindowsCommandExecutor';
+        }
+
+        $this->commandExecutor = new $executorClass(
+            $this->buildLogger,
+            PHPCI_DIR,
+            $this->quiet,
+            $this->verbose
+        );
+
+        $this->interpolator = new BuildInterpolator();
+    }
+
+    /**
+     * Set the config array, as read from phpci.yml
+     *
+     * @param array|null $config
+     *
+     * @throws Exception
+     */
+    public function setConfigArray($config)
+    {
+        if (is_null($config) || !is_array($config)) {
+            throw new Exception(Lang::get('missing_phpci_yml'));
+        }
+
+        $this->config = $config;
+    }
+
+    /**
+     * Access a variable from the phpci.yml file.
+     *
+     * @param string
+     *
+     * @return mixed
+     */
+    public function getConfig($key)
+    {
+        $rtn = null;
+
+        if (isset($this->config[$key])) {
+            $rtn = $this->config[$key];
+        }
+
+        return $rtn;
+    }
+
+    /**
+     * Access a variable from the config.yml
+     *
+     * @param $key
+     *
+     * @return mixed
+     */
+    public function getSystemConfig($key)
+    {
+        return Config::getInstance()->get($key);
+    }
+
+    /**
+     * Run the active build.
+     */
+    public function execute()
+    {
+        // Update the build in the database, ping any external services.
+        $this->build->setStatus(Build::STATUS_RUNNING);
+        $this->build->setStarted(new DateTime());
+        $this->store->save($this->build);
+        $this->build->sendStatusPostback();
+        $success = true;
+
+        try {
+            // Set up the build:
+            $this->setupBuild();
+
+            // Run the core plugin stages:
+            foreach (array('setup', 'test') as $stage) {
+                $success &= $this->pluginExecutor->executePlugins($this->config, $stage);
+            }
+
+            // Set the status so this can be used by complete, success and failure
+            // stages.
+            if ($success) {
+                $this->build->setStatus(Build::STATUS_SUCCESS);
+            } else {
+                $this->build->setStatus(Build::STATUS_FAILED);
+            }
+
+            // Complete stage plugins are always run
+            $this->pluginExecutor->executePlugins($this->config, 'complete');
+
+            if ($success) {
+                $this->pluginExecutor->executePlugins($this->config, 'success');
+                $this->buildLogger->logSuccess(Lang::get('build_success'));
+            } else {
+                $this->pluginExecutor->executePlugins($this->config, 'failure');
+                $this->buildLogger->logFailure(Lang::get('build_failed'));
+            }
+        } catch (Exception $ex) {
+            $this->build->setStatus(Build::STATUS_FAILED);
+            $this->buildLogger->logFailure(Lang::get('exception') . $ex->getMessage());
+        }
+
+        // Update the build in the database, ping any external services, etc.
+        $this->build->sendStatusPostback();
+        $this->build->setFinished(new DateTime());
+        $this->store->save($this->build);
+        $this->removeBuildDirectory();
+    }
+
+    /**
+     * Set up a working copy of the project for building.
+     */
+    protected function setupBuild()
+    {
+        $this->buildPath = PHPCI_DIR . 'PHPCI/build/' . $this->build->getId() . '/';
+        $this->build->currentBuildPath = $this->buildPath;
+
+        $this->interpolator->setupInterpolationVars(
+            $this->build,
+            $this->buildPath,
+            PHPCI_URL
+        );
+
+        $this->commandExecutor->setBuildPath($this->buildPath);
+
+        // Create a working copy of the project:
+        if (!$this->build->createWorkingCopy($this, $this->buildPath)) {
+            throw new Exception(Lang::get('could_not_create_working'));
+        }
+
+        // Does the project's phpci.yml request verbose mode?
+        if (!isset($this->config['build_settings']['verbose']) || !$this->config['build_settings']['verbose']) {
+            $this->verbose = false;
+        }
+
+        // Does the project have any paths it wants plugins to ignore?
+        if (isset($this->config['build_settings']['ignore'])) {
+            $this->ignore = $this->config['build_settings']['ignore'];
+        }
+
+        $this->buildLogger->logSuccess(Lang::get('working_copy_created', $this->buildPath));
+        return true;
+    }
+
+    /**
+     * Returns a configured instance of the plugin factory.
+     *
+     * @param Build $build
+     *
+     * @return PluginFactory
+     */
+    private function buildPluginFactory(Build $build)
+    {
+        $pluginFactory = new PluginFactory();
+
+        $self = $this;
+        $pluginFactory->registerResource(
+            function () use ($self) {
+                return $self;
+            },
+            null,
+            'PHPCI\Builder'
+        );
+
+        $pluginFactory->registerResource(
+            function () use ($build) {
+                return $build;
+            },
+            null,
+            'PHPCI\Model\Build'
+        );
+
+        $logger = $this->logger;
+        $pluginFactory->registerResource(
+            function () use ($logger) {
+                return $logger;
+            },
+            null,
+            'Psr\Log\LoggerInterface'
+        );
+
+        $pluginFactory->registerResource(
+            function () use ($self) {
+                $factory = new MailerFactory($self->getSystemConfig('phpci'));
+                return $factory->getSwiftMailerFromConfig();
+            },
+            null,
+            'Swift_Mailer'
+        );
+
+        return $pluginFactory;
+    }
+
+    /**
+     * Sets a logger instance on the object
+     *
+     * @param LoggerInterface $logger
+     *
+     * @return null
+     */
+    public function setLogger(LoggerInterface $logger)
+    {
+        $this->buildLogger->setLogger($logger);
+    }
+
+    /**
+     * Remove the build directory of a finished build.
+     */
+    public function removeBuildDirectory()
+    {
+        if (!is_dir($this->buildPath)) {
+            return;
+        }
+
+        $this->buildLogger->log(Lang::get('removing_build'));
+
+        $cmd = 'rm -Rf "%s"';
+        if (IS_WIN) {
+            $cmd = 'rmdir /S /Q "%s"';
+        }
+        $this->commandExecutor->executeCommand(array($cmd, rtrim($this->buildPath, '/')));
+    }
+}
